% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKHR52.R
\name{PKHR52}
\alias{PKHR52}
\title{Pakistan Demographic and Health Survey 2006-07 Household Recode Data}
\description{
\code{PKHR52} data of Household Recode from Pakistan Demographic and Health Survey 2006-07.
\describe{
    \item{\code{hhid}}{case identification}                                          
    \item{\code{hv000}}{country code and phase}                                      
    \item{\code{hv001}}{cluster number}                                              
    \item{\code{hv002}}{household number}                                            
    \item{\code{hv003}}{respondent's line number}                                    
    \item{\code{hv004}}{ultimate area unit}                                          
    \item{\code{hv005}}{sample weight}                                               
    \item{\code{hv006}}{month of interview}                                          
    \item{\code{hv007}}{year of interview}                                           
    \item{\code{hv008}}{date of interview (cmc)}                                     
    \item{\code{hv009}}{number of household members}                                 
    \item{\code{hv010}}{number of eligible women in hh}                              
    \item{\code{hv011}}{number of eligible men in hh}                                
    \item{\code{hv012}}{number of de jure members}                                   
    \item{\code{hv013}}{number of de facto members}                                  
    \item{\code{hv014}}{number of children 5 and under}                              
    \item{\code{hv015}}{result of household interview}                               
    \item{\code{hv016}}{day of interview}                                            
    \item{\code{hv017}}{number of visits}                                            
    \item{\code{hv018}}{interviewer identification}                                  
    \item{\code{hv019}}{keyer identification}                                        
    \item{\code{hv020}}{ever-married sample}                                         
    \item{\code{hv021}}{primary sampling unit}                                       
    \item{\code{hv022}}{sample stratum number}                                       
    \item{\code{hv023}}{sample domain}                                               
    \item{\code{hv024}}{region}                                                      
    \item{\code{hv025}}{type of place of residence}                                  
    \item{\code{hv026}}{place of residence}                                          
    \item{\code{hv027}}{selection for male/husb. int.}                               
    \item{\code{hv028}}{sample weight for male subsamp}                              
    \item{\code{hv030}}{field supervisor}                                            
    \item{\code{hv031}}{field editor}                                                
    \item{\code{hv032}}{office editor}                                               
    \item{\code{hv033}}{ultimate area selection pr-na}                               
    \item{\code{hv035}}{number of eligible children for height & weight}             
    \item{\code{hv040}}{cluster altitude in meters -na}                              
    \item{\code{hv041}}{total adults measured}                                       
    \item{\code{hv042}}{selection for hemoglobin}                                    
    \item{\code{hv043}}{selection for women's statu-na}                              
    \item{\code{hv044}}{selection for domestic viol-na}                              
    \item{\code{hv201}}{source of drinking water}                                    
    \item{\code{hv202}}{source of non-drinking wate-na}                              
    \item{\code{hv204}}{time to get to water source}                                 
    \item{\code{hv205}}{type of toilet facility}                                     
    \item{\code{hv206}}{has electricity}                                             
    \item{\code{hv207}}{has radio}                                                   
    \item{\code{hv208}}{has television}                                              
    \item{\code{hv209}}{has refrigerator}                                            
    \item{\code{hv210}}{has bicycle}                                                 
    \item{\code{hv211}}{has motorcycle/scooter}                                      
    \item{\code{hv212}}{has car/truck}                                               
    \item{\code{hv213}}{main floor material}                                         
    \item{\code{hv214}}{main wall material}                                          
    \item{\code{hv215}}{main roof material}                                          
    \item{\code{hv216}}{rooms used for sleeping}                                     
    \item{\code{hv217}}{relationship structure}                                      
    \item{\code{hv218}}{line number of head of househ.}                              
    \item{\code{hv219}}{sex of head of household}                                    
    \item{\code{hv220}}{age of head of household}                                    
    \item{\code{hv221}}{has telephone}                                               
    \item{\code{hv225}}{share toilet with other households}                          
    \item{\code{hv226}}{type of cooking fuel}                                        
    \item{\code{hv227}}{have bednet for sleeping}                                    
    \item{\code{hv228}}{children under 5 slept under bednet last night}              
    \item{\code{hv234}}{test salt for iodine       -na}                              
    \item{\code{hv235}}{location of source for wate-na}                              
    \item{\code{hv236}}{person fetching water      -na}                              
    \item{\code{hv237}}{anything done to water to make safe to drink}                
    \item{\code{hv237a}}{water usually treated by: boil}                             
    \item{\code{hv237b}}{water usually treated by: add bleach/chlorine}              
    \item{\code{hv237c}}{water usually treated by: strain through a cloth}           
    \item{\code{hv237d}}{water usually treated by: use water filter}                 
    \item{\code{hv237e}}{water usually treated by: solar disinfection}               
    \item{\code{hv237f}}{water usually treated by: let it stand and settle}          
    \item{\code{hv237g}}{water usually treated by: c-na}                             
    \item{\code{hv237h}}{water usually treated by: c-na}                             
    \item{\code{hv237i}}{water usually treated by: c-na}                             
    \item{\code{hv237j}}{water usually treated by: c-na}                             
    \item{\code{hv237k}}{water usually treated by: c-na}                             
    \item{\code{hv237x}}{water usually treated by: other}                            
    \item{\code{hv237z}}{water usually treated by: don't know}                       
    \item{\code{hv238}}{number of households sharin-na}                              
    \item{\code{hv239}}{food cooked on stove or ope-na}                              
    \item{\code{hv240}}{household has a chimney, ho-na}                              
    \item{\code{hv241}}{food cooked in the house / -na}                              
    \item{\code{hv242}}{household has separate room-na}                              
    \item{\code{hv243a}}{has a mobile telephone     -na}                             
    \item{\code{hv243b}}{has a watch}                                                
    \item{\code{hv243c}}{has an animal-drawn cart}                                   
    \item{\code{hv243d}}{has a boat with a motor}                                    
    \item{\code{hv244}}{own land usable for agriculture}                             
    \item{\code{hv245}}{hectares for agricultural l-na}                              
    \item{\code{hv246}}{livestock, herds or farm animals}                            
    \item{\code{hv246a}}{cattle own                 -na}                             
    \item{\code{hv246b}}{cows, bulls own}                                            
    \item{\code{hv246c}}{horses, donkeys, mules own}                                 
    \item{\code{hv246d}}{goats own}                                                  
    \item{\code{hv246e}}{sheep own}                                                  
    \item{\code{hv246f}}{chickens own}                                               
    \item{\code{hv246g}}{buffalo own}                                                
    \item{\code{hv246h}}{camels own}                                                 
    \item{\code{hv246i}}{cs own                     -na}                             
    \item{\code{hv246j}}{cs own                     -na}                             
    \item{\code{hv246k}}{cs own                     -na}                             
    \item{\code{hv247}}{owns a bank account        -na}                              
    \item{\code{hv248}}{number of sick people 18-59-na}                              
    \item{\code{hv249}}{member of the hh died last -na}                              
    \item{\code{hv250}}{number of members who died -na}                              
    \item{\code{hv251}}{number of orphans and vulne-na}                              
    \item{\code{hv270}}{wealth index}                                                
    \item{\code{hv271}}{wealth index factor score (5 decimals)}                      
    \item{\code{hml1}}{number of mosquito nets}                                      
    \item{\code{hml1a}}{number of mosquito nets with specific information}           
    \item{\code{hml2}}{number of children under bednet previous night}               
    \item{\code{shdist}}{district}                                                   
    \item{\code{shtype}}{type of questionnaire}                                      
    \item{\code{shv005}}{hh weight for all households}                               
    \item{\code{shv218}}{line number of head of household (all households).}         
    \item{\code{shv219}}{sex of head of household (all households)}                  
    \item{\code{shv220}}{age of head of household (all households)}                  
    \item{\code{shbirths}}{total births in household since january 2003}             
    \item{\code{shdeaths}}{total deaths in household since january 2003}             
    \item{\code{shcdeath}}{total deaths < 5/sbs from q38}                            
    \item{\code{shfdeath}}{total female deaths age 12-49 from q39}                   
    \item{\code{sh107f}}{room cooler, air conditioner owns}                          
    \item{\code{sh107g}}{washing machine owns}                                       
    \item{\code{sh107h}}{water pump owns}                                            
    \item{\code{sh107i}}{bed owns}                                                   
    \item{\code{sh107j}}{chairs owns}                                                
    \item{\code{sh107k}}{almirah/cabinet owns}                                       
    \item{\code{sh107l}}{clock owns}                                                 
    \item{\code{sh107m}}{sofa owns}                                                  
    \item{\code{sh107n}}{sewing machine owns}                                        
    \item{\code{sh107o}}{camera owns}                                                
    \item{\code{sh107p}}{personal computer owns}                                     
    \item{\code{sh113}}{status of house}                                             
    \item{\code{sh127a}}{hh uses coil to avoid mosquitos}                            
    \item{\code{sh127b}}{hh uses mats to avoid mosquitos}                            
    \item{\code{sh127c}}{hh uses spray to avoid mosquitos}                           
    \item{\code{sh127d}}{hh uses electric spray repellant to avoid mosquitos}        
    \item{\code{sh127e}}{hh uses insect repellant to avoid mosquitos}                
    \item{\code{sh127f}}{hh uses smoke to avoid mosquitos}                           
    \item{\code{sh127g}}{hh uses fan to avoid mosquitos}                             
    \item{\code{sh127h}}{hh uses covered windows/screens to avoid mosquitos}         
    \item{\code{sh127x}}{hh uses other to avoid mosquitos}                           
    \item{\code{sh127y}}{hh uses nothing to avoid mosquitos}                         
    \item{\code{sh128}}{currently have any medicines for treating malaria in the hou}
    \item{\code{hmlidx.1}}{net number}                                               
    \item{\code{hmlidx.2}}{net number}                                               
    \item{\code{hml3.1}}{net observed               -na}                             
    \item{\code{hml3.2}}{net observed               -na}                             
    \item{\code{hml4.1}}{months ago obtained mosquit-na}                             
    \item{\code{hml4.2}}{months ago obtained mosquit-na}                             
    \item{\code{hml5.1}}{was net treated with insecticide when bought}               
    \item{\code{hml5.2}}{was net treated with insecticide when bought}               
    \item{\code{hml6.1}}{mosquito net treated with i-na}                             
    \item{\code{hml6.2}}{mosquito net treated with i-na}                             
    \item{\code{hml7.1}}{brand of bednet            -na}                             
    \item{\code{hml7.2}}{brand of bednet            -na}                             
    \item{\code{hml8.1}}{bednet treated since receiving}                             
    \item{\code{hml8.2}}{bednet treated since receiving}                             
    \item{\code{hml9.1}}{months ago re-treated}                                      
    \item{\code{hml9.2}}{months ago re-treated}                                      
    \item{\code{hml10.1}}{itn net}                                                   
    \item{\code{hml10.2}}{itn net}                                                   
    \item{\code{hml11.1}}{number of persons who slept under this net}                
    \item{\code{hml11.2}}{number of persons who slept under this net}                
    \item{\code{hml21.1}}{did someone sleep under bednet last night}                 
    \item{\code{hml21.2}}{did someone sleep under bednet last night}                 
    \item{\code{hmla.1}}{person who slept in mosquito net}                           
    \item{\code{hmla.2}}{person who slept in mosquito net}                           
    \item{\code{hmlb.1}}{person who slept in mosquito net}                           
    \item{\code{hmlb.2}}{person who slept in mosquito net}                           
    \item{\code{hmlc.1}}{person who slept in mosquito net}                           
    \item{\code{hmlc.2}}{person who slept in mosquito net}                           
    \item{\code{hmld.1}}{person who slept in mosquit-na}                             
    \item{\code{hmld.2}}{person who slept in mosquit-na}                             
    \item{\code{hmle.1}}{person who slept in mosquit-na}                             
    \item{\code{hmle.2}}{person who slept in mosquit-na}                             
    \item{\code{sh20.1}}{line number}                                                
    \item{\code{sh20.2}}{line number}                                                
    \item{\code{sh20.3}}{line number}                                                
    \item{\code{sh20.4}}{line number}                                                
    \item{\code{sh20.5}}{line number}                                                
    \item{\code{sh20.6}}{line number}                                                
    \item{\code{sh20.7}}{line number}                                                
    \item{\code{sh20.8}}{line number}                                                
    \item{\code{sh20.9}}{line number}                                                
    \item{\code{sh22.1}}{gender}                                                     
    \item{\code{sh22.2}}{gender}                                                     
    \item{\code{sh22.3}}{gender}                                                     
    \item{\code{sh22.4}}{gender}                                                     
    \item{\code{sh22.5}}{gender}                                                     
    \item{\code{sh22.6}}{gender}                                                     
    \item{\code{sh22.7}}{gender}                                                     
    \item{\code{sh22.8}}{gender}                                                     
    \item{\code{sh22.9}}{gender}                                                     
    \item{\code{sh23m.1}}{month of birth}                                            
    \item{\code{sh23m.2}}{month of birth}                                            
    \item{\code{sh23m.3}}{month of birth}                                            
    \item{\code{sh23m.4}}{month of birth}                                            
    \item{\code{sh23m.5}}{month of birth}                                            
    \item{\code{sh23m.6}}{month of birth}                                            
    \item{\code{sh23m.7}}{month of birth}                                            
    \item{\code{sh23m.8}}{month of birth}                                            
    \item{\code{sh23m.9}}{month of birth}                                            
    \item{\code{sh23y.1}}{year of birth}                                             
    \item{\code{sh23y.2}}{year of birth}                                             
    \item{\code{sh23y.3}}{year of birth}                                             
    \item{\code{sh23y.4}}{year of birth}                                             
    \item{\code{sh23y.5}}{year of birth}                                             
    \item{\code{sh23y.6}}{year of birth}                                             
    \item{\code{sh23y.7}}{year of birth}                                             
    \item{\code{sh23y.8}}{year of birth}                                             
    \item{\code{sh23y.9}}{year of birth}                                             
    \item{\code{sh24.1}}{live birth}                                                 
    \item{\code{sh24.2}}{live birth}                                                 
    \item{\code{sh24.3}}{live birth}                                                 
    \item{\code{sh24.4}}{live birth}                                                 
    \item{\code{sh24.5}}{live birth}                                                 
    \item{\code{sh24.6}}{live birth}                                                 
    \item{\code{sh24.7}}{live birth}                                                 
    \item{\code{sh24.8}}{live birth}                                                 
    \item{\code{sh24.9}}{live birth}                                                 
    \item{\code{sh25.1}}{currently alive}                                            
    \item{\code{sh25.2}}{currently alive}                                            
    \item{\code{sh25.3}}{currently alive}                                            
    \item{\code{sh25.4}}{currently alive}                                            
    \item{\code{sh25.5}}{currently alive}                                            
    \item{\code{sh25.6}}{currently alive}                                            
    \item{\code{sh25.7}}{currently alive}                                            
    \item{\code{sh25.8}}{currently alive}                                            
    \item{\code{sh25.9}}{currently alive}                                            
    \item{\code{sh26.1}}{line number in household roster}                            
    \item{\code{sh26.2}}{line number in household roster}                            
    \item{\code{sh26.3}}{line number in household roster}                            
    \item{\code{sh26.4}}{line number in household roster}                            
    \item{\code{sh26.5}}{line number in household roster}                            
    \item{\code{sh26.6}}{line number in household roster}                            
    \item{\code{sh26.7}}{line number in household roster}                            
    \item{\code{sh26.8}}{line number in household roster}                            
    \item{\code{sh26.9}}{line number in household roster}                            
    \item{\code{sh23c.1}}{cmc birth}                                                 
    \item{\code{sh23c.2}}{cmc birth}                                                 
    \item{\code{sh23c.3}}{cmc birth}                                                 
    \item{\code{sh23c.4}}{cmc birth}                                                 
    \item{\code{sh23c.5}}{cmc birth}                                                 
    \item{\code{sh23c.6}}{cmc birth}                                                 
    \item{\code{sh23c.7}}{cmc birth}                                                 
    \item{\code{sh23c.8}}{cmc birth}                                                 
    \item{\code{sh23c.9}}{cmc birth}                                                 
    \item{\code{sh23f.1}}{cmc flag}                                                  
    \item{\code{sh23f.2}}{cmc flag}                                                  
    \item{\code{sh23f.3}}{cmc flag}                                                  
    \item{\code{sh23f.4}}{cmc flag}                                                  
    \item{\code{sh23f.5}}{cmc flag}                                                  
    \item{\code{sh23f.6}}{cmc flag}                                                  
    \item{\code{sh23f.7}}{cmc flag}                                                  
    \item{\code{sh23f.8}}{cmc flag}                                                  
    \item{\code{sh23f.9}}{cmc flag}                                                  
    \item{\code{sh29.1}}{line number}                                                
    \item{\code{sh29.2}}{line number}                                                
    \item{\code{sh29.3}}{line number}                                                
    \item{\code{sh29.4}}{line number}                                                
    \item{\code{sh29.5}}{line number}                                                
    \item{\code{sh29.6}}{line number}                                                
    \item{\code{sh31a.1}}{usual resident or visitor}                                 
    \item{\code{sh31a.2}}{usual resident or visitor}                                 
    \item{\code{sh31a.3}}{usual resident or visitor}                                 
    \item{\code{sh31a.4}}{usual resident or visitor}                                 
    \item{\code{sh31a.5}}{usual resident or visitor}                                 
    \item{\code{sh31a.6}}{usual resident or visitor}                                 
    \item{\code{sh31.1}}{gender}                                                     
    \item{\code{sh31.2}}{gender}                                                     
    \item{\code{sh31.3}}{gender}                                                     
    \item{\code{sh31.4}}{gender}                                                     
    \item{\code{sh31.5}}{gender}                                                     
    \item{\code{sh31.6}}{gender}                                                     
    \item{\code{sh32m.1}}{month of death}                                            
    \item{\code{sh32m.2}}{month of death}                                            
    \item{\code{sh32m.3}}{month of death}                                            
    \item{\code{sh32m.4}}{month of death}                                            
    \item{\code{sh32m.5}}{month of death}                                            
    \item{\code{sh32m.6}}{month of death}                                            
    \item{\code{sh32y.1}}{year of death}                                             
    \item{\code{sh32y.2}}{year of death}                                             
    \item{\code{sh32y.3}}{year of death}                                             
    \item{\code{sh32y.4}}{year of death}                                             
    \item{\code{sh32y.5}}{year of death}                                             
    \item{\code{sh32y.6}}{year of death}                                             
    \item{\code{sh33.1}}{age at death}                                               
    \item{\code{sh33.2}}{age at death}                                               
    \item{\code{sh33.3}}{age at death}                                               
    \item{\code{sh33.4}}{age at death}                                               
    \item{\code{sh33.5}}{age at death}                                               
    \item{\code{sh33.6}}{age at death}                                               
    \item{\code{sh34.1}}{death was female, 12-49 years of age}                       
    \item{\code{sh34.2}}{death was female, 12-49 years of age}                       
    \item{\code{sh34.3}}{death was female, 12-49 years of age}                       
    \item{\code{sh34.4}}{death was female, 12-49 years of age}                       
    \item{\code{sh34.5}}{death was female, 12-49 years of age}                       
    \item{\code{sh34.6}}{death was female, 12-49 years of age}                       
    \item{\code{sh35.1}}{pregnant at the time of death}                              
    \item{\code{sh35.2}}{pregnant at the time of death}                              
    \item{\code{sh35.3}}{pregnant at the time of death}                              
    \item{\code{sh35.4}}{pregnant at the time of death}                              
    \item{\code{sh35.5}}{pregnant at the time of death}                              
    \item{\code{sh35.6}}{pregnant at the time of death}                              
    \item{\code{sh36.1}}{died during childbirth}                                     
    \item{\code{sh36.2}}{died during childbirth}                                     
    \item{\code{sh36.3}}{died during childbirth}                                     
    \item{\code{sh36.4}}{died during childbirth}                                     
    \item{\code{sh36.5}}{died during childbirth}                                     
    \item{\code{sh36.6}}{died during childbirth}                                     
    \item{\code{sh37.1}}{died within 6-weeks after delivery}                         
    \item{\code{sh37.2}}{died within 6-weeks after delivery}                         
    \item{\code{sh37.3}}{died within 6-weeks after delivery}                         
    \item{\code{sh37.4}}{died within 6-weeks after delivery}                         
    \item{\code{sh37.5}}{died within 6-weeks after delivery}                         
    \item{\code{sh37.6}}{died within 6-weeks after delivery}                         
    \item{\code{sh32c.1}}{cmc death}                                                 
    \item{\code{sh32c.2}}{cmc death}                                                 
    \item{\code{sh32c.3}}{cmc death}                                                 
    \item{\code{sh32c.4}}{cmc death}                                                 
    \item{\code{sh32c.5}}{cmc death}                                                 
    \item{\code{sh32c.6}}{cmc death}                                                 
    \item{\code{sh32f.1}}{cmc flag}                                                  
    \item{\code{sh32f.2}}{cmc flag}                                                  
    \item{\code{sh32f.3}}{cmc flag}                                                  
    \item{\code{sh32f.4}}{cmc flag}                                                  
    \item{\code{sh32f.5}}{cmc flag}                                                  
    \item{\code{sh32f.6}}{cmc flag}                                                  


  }
}
\examples{
data("PKHR52")
PKHR52
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
