% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKHW21.R
\name{PKHW21}
\alias{PKHW21}
\title{Pakistan Demographic and Health Survey 1990-91 Height and Weight of Children Data}
\description{
\code{PKHW21} data of Height and Weight of Children from Pakistan Demographic and Health Survey 1990-91.
\describe{
    \item{\code{hwcaseid}}{case identification}                     
    \item{\code{hwline}}{line}                                      
    \item{\code{hwlevel}}{data from hh or individual}               
    \item{\code{hc70}}{ht/a standard deviations (according to who)} 
    \item{\code{hc71}}{wt/a standard deviations (according to who)} 
    \item{\code{hc72}}{wt/ht standard deviations (according to who)}
    \item{\code{hc73}}{bmi standard deviations (according to who)}  

  }
}
\examples{
data("PKHW21")
PKHW21
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
