% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPR21.R
\name{PKPR21}
\alias{PKPR21}
\title{Pakistan Demographic and Health Survey 1990-91 Household Member Recode Data}
\description{
\code{PKPR21} data of Household Member Recode from Pakistan Demographic and Health Survey 1990-91.
\describe{
    \item{\code{hhid}}{case identification}            
    \item{\code{hv000}}{country code and phase}        
    \item{\code{hv001}}{cluster number}                
    \item{\code{hv002}}{household number}              
    \item{\code{hv003}}{respondent's line number}      
    \item{\code{hv004}}{ultimate area unit}            
    \item{\code{hv005}}{sample weight}                 
    \item{\code{hv006}}{month of interview}            
    \item{\code{hv007}}{year of interview}             
    \item{\code{hv008}}{date of interview (cmc)}       
    \item{\code{hv009}}{number of household members}   
    \item{\code{hv010}}{number of eligible women in hh}
    \item{\code{hv011}}{number of eligible men in hh}  
    \item{\code{hv012}}{number of de jure members}     
    \item{\code{hv013}}{number of de facto members}    
    \item{\code{hv014}}{number of children 5 and under}
    \item{\code{hv015}}{result of household interview} 
    \item{\code{hv016}}{day of interview}              
    \item{\code{hv017}}{number of visits}              
    \item{\code{hv018}}{interviewer identification}    
    \item{\code{hv019}}{keyer identification}          
    \item{\code{hv020}}{ever-married sample}           
    \item{\code{hv021}}{primary sampling unit}         
    \item{\code{hv022}}{sample stratum number}         
    \item{\code{hv023}}{sample domain}                 
    \item{\code{hv024}}{region}                        
    \item{\code{hv025}}{type of place of residence}    
    \item{\code{hv026}}{place of residence}            
    \item{\code{hv027}}{selection for male/husb. int.} 
    \item{\code{hv028}}{sample weight for male subsamp}
    \item{\code{hv201}}{source of drinking water}      
    \item{\code{hv202}}{source of non-drinking water}  
    \item{\code{hv203}}{same source of water}          
    \item{\code{hv204}}{time to get to water source}   
    \item{\code{hv205}}{type of toilet facility}       
    \item{\code{hv206}}{has electricity}               
    \item{\code{hv207}}{has radio}                     
    \item{\code{hv208}}{has television}                
    \item{\code{hv209}}{has refrigerator}              
    \item{\code{hv210}}{has bicycle}                   
    \item{\code{hv211}}{has motorcycle}                
    \item{\code{hv212}}{has car                   - na}
    \item{\code{hv213}}{main floor material       - na}
    \item{\code{hv214}}{main wall material}            
    \item{\code{hv215}}{main roof material}            
    \item{\code{hv216}}{rooms for sleeping}            
    \item{\code{hv217}}{relationship structure}        
    \item{\code{hv218}}{line number of head of househ.}
    \item{\code{hv219}}{sex of head of household}      
    \item{\code{hv220}}{age of head of household}      
    \item{\code{shdist}}{district}                     
    \item{\code{shclust}}{cluster number}              
    \item{\code{sh6e}}{has a room cooler}              
    \item{\code{sh6f}}{has a washing machine}          
    \item{\code{sh6g}}{has a water pump}               
    \item{\code{sh7c}}{has a car, van or tractor}      
    \item{\code{hvidx}}{line number}                   
    \item{\code{hv101}}{relationship to head}          
    \item{\code{hv102}}{usual resident}                
    \item{\code{hv103}}{slept last night}              
    \item{\code{hv104}}{sex of household member}       
    \item{\code{hv105}}{age of household members}      
    \item{\code{hv106}}{highest educational level}     
    \item{\code{hv107}}{highest year of education}     
    \item{\code{hv108}}{education in single years}     
    \item{\code{hv109}}{educational attainment}        
    \item{\code{hv110}}{member still in school}        
    \item{\code{hv111}}{mother alive              - na}
    \item{\code{hv112}}{mother's line number}          
    \item{\code{hv113}}{father alive              - na}
    \item{\code{hv114}}{father's line number}          
    \item{\code{hv115}}{current marital status}        
    \item{\code{hv116}}{currently, formerly, never m.} 
    \item{\code{hv117}}{eligibility for female intview}
    \item{\code{hv118}}{eligibility for male interview}
    \item{\code{idxh4}}{index to household schedule}   
    \item{\code{shlevel}}{level of education}          
    \item{\code{shgrade}}{highest grade completed}     
    \item{\code{hhid}}{case identification}            
    \item{\code{hv000}}{country code and phase}        
    \item{\code{hv001}}{cluster number}                
    \item{\code{hv002}}{household number}              
    \item{\code{hv003}}{respondent's line number}      
    \item{\code{hv004}}{ultimate area unit}            
    \item{\code{hv005}}{sample weight}                 
    \item{\code{hv006}}{month of interview}            
    \item{\code{hv007}}{year of interview}             
    \item{\code{hv008}}{date of interview (cmc)}       
    \item{\code{hv009}}{number of household members}   
    \item{\code{hv010}}{number of eligible women in hh}
    \item{\code{hv011}}{number of eligible men in hh}  
    \item{\code{hv012}}{number of de jure members}     
    \item{\code{hv013}}{number of de facto members}    
    \item{\code{hv014}}{number of children 5 and under}
    \item{\code{hv015}}{result of household interview} 
    \item{\code{hv016}}{day of interview}              
    \item{\code{hv017}}{number of visits}              
    \item{\code{hv018}}{interviewer identification}    
    \item{\code{hv019}}{keyer identification}          
    \item{\code{hv020}}{ever-married sample}           
    \item{\code{hv021}}{primary sampling unit}         
    \item{\code{hv022}}{sample stratum number}         
    \item{\code{hv023}}{sample domain}                 
    \item{\code{hv024}}{region}                        
    \item{\code{hv025}}{type of place of residence}    
    \item{\code{hv026}}{place of residence}            
    \item{\code{hv027}}{selection for male/husb. int.} 
    \item{\code{hv028}}{sample weight for male subsamp}
    \item{\code{hv201}}{source of drinking water}      
    \item{\code{hv202}}{source of non-drinking water}  
    \item{\code{hv203}}{same source of water}          
    \item{\code{hv204}}{time to get to water source}   
    \item{\code{hv205}}{type of toilet facility}       
    \item{\code{hv206}}{has electricity}               
    \item{\code{hv207}}{has radio}                     
    \item{\code{hv208}}{has television}                
    \item{\code{hv209}}{has refrigerator}              
    \item{\code{hv210}}{has bicycle}                   
    \item{\code{hv211}}{has motorcycle}                
    \item{\code{hv212}}{has car                   - na}
    \item{\code{hv213}}{main floor material       - na}
    \item{\code{hv214}}{main wall material}            
    \item{\code{hv215}}{main roof material}            
    \item{\code{hv216}}{rooms for sleeping}            
    \item{\code{hv217}}{relationship structure}        
    \item{\code{hv218}}{line number of head of househ.}
    \item{\code{hv219}}{sex of head of household}      
    \item{\code{hv220}}{age of head of household}      
    \item{\code{shdist}}{district}                     
    \item{\code{shclust}}{cluster number}              
    \item{\code{sh6e}}{has a room cooler}              
    \item{\code{sh6f}}{has a washing machine}          
    \item{\code{sh6g}}{has a water pump}               
    \item{\code{sh7c}}{has a car, van or tractor}      
    \item{\code{hvidx}}{line number}                   
    \item{\code{hv101}}{relationship to head}          
    \item{\code{hv102}}{usual resident}                
    \item{\code{hv103}}{slept last night}              
    \item{\code{hv104}}{sex of household member}       
    \item{\code{hv105}}{age of household members}      
    \item{\code{hv106}}{highest educational level}     
    \item{\code{hv107}}{highest year of education}     
    \item{\code{hv108}}{education in single years}     
    \item{\code{hv109}}{educational attainment}        
    \item{\code{hv110}}{member still in school}        
    \item{\code{hv111}}{mother alive              - na}
    \item{\code{hv112}}{mother's line number}          
    \item{\code{hv113}}{father alive              - na}
    \item{\code{hv114}}{father's line number}          
    \item{\code{hv115}}{current marital status}        
    \item{\code{hv116}}{currently, formerly, never m.} 
    \item{\code{hv117}}{eligibility for female intview}
    \item{\code{hv118}}{eligibility for male interview}
    \item{\code{idxh4}}{index to household schedule}   
    \item{\code{shlevel}}{level of education}          
    \item{\code{shgrade}}{highest grade completed}     


  }
}
\examples{
load("PKPR21")
PKPR21
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
