% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKWI21.R
\name{PKWI21}
\alias{PKWI21}
\title{Pakistan Demographic and Health Survey 1990-91 Wealth Index Data}
\description{
\code{PKWI21} data of Wealth Index from Pakistan Demographic and Health Survey 1990-91.
\describe{
    \item{\code{whhid}}{household id}                      
    \item{\code{wlthindf}}{wealth index from factor scores}
    \item{\code{wlthind5}}{quintiles of wealth index}      
    \item{\code{whhid}}{household id}                      
    \item{\code{wlthindf}}{wealth index from factor scores}
    \item{\code{wlthind5}}{quintiles of wealth index}      

  }
}
\examples{
load("PKWI21")
PKWI21
}
\references{
\enumerate{
  \item National Institute of Population Studies (NIPS), Pakistan (\url{http://nips.org.pk/Home.htm}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  }
}
